services/radxa3w_camera/requirements.txt
flask
flask-cors
gunicorn
requests
opencv-python
numpy
pillow